First pass on updating interpolant-based mc
Using the regular solver as the main one (for compliance with the Prover class)
This is mostly straightforward, unfortunately there are two ugly changes:

1. We now need two translators, because we need to translate terms to the interpolator, but also back to the solver (e.g. the interpolant itself)
2. Since symbols are already created in the term_translator, we need to populate the cache for it's translator. I noticed a weird corner case in the unroller doing this.

If you ask for a symbol at_time(s, 1) before getting the symbol at 0, then it won't untime correctly later. I think this is because the current and next-state variables are
not assigned correctly.

Note: this also required updating the smt-switch version to get upstream changes