Fix test path to be relative to file
If you run `pytest tests` from the project root, then this test will
always pass since the `coreir_inputs_directory` will be empty (it
expects the encoders directory to be relative to the test invocation
path, so it won't find it properly).

This updates the test to use a path that's explictly relative to the
test file's location.