Separate COI analysis from TS transformation (#158)
* Fix contributors

* Clarify that ConeOfInfluence is StaticConeOfInfluence modifier

* Get rid of collect_coi_term -- can just insert into a set

* Factor out functional cone of influence analysis from the static
COI transformation pass. E.g. we might want to check COI without
modifying a transition system.

Also remove support for to_remove in cone-of-influence. It is hacky
and won't do what people expect if the input format is BTOR2. If
there was complex term in to_remove (for example obtained via the
wirename) this will stop COI but it might not be the only place
symbols of interest appear in the the terms, because Yosys->BTOR2
loses some of the circuit structure.

* Fix python bindings