Use label in more places
Have issue that not all ic3 versions use it the same

However, this might end up being irrelevant once this is cleaned up
more. Because inductive_generalization should be implemented by
the backend anyway.

For now just want to avoid errors.

Note: there is an issue with boolector where it will substitute.
For example say you have symbol .pred2 which stands for
(= counter #b0000000)

but then you end up creating (not .pred2) this will instead return
(not (= counter #b0000000))

by substitution because of the top-level assertion