Fixes in IC3SA (#238)
Message:
This PR adds a fallback procedure for functional refinement. There's an issue where functional refinement can result in a value due to simplification in the smt-solver. In that case, it will fall back to a simpler refinement approach. Additionally, there are cases (especially when using actual input variables that cannot have lemmas learned over them) where functional refinement can add the same lemma again due to simplification. Another branch (ic3sa-fixes) is experimenting with using a logging solver to avoid that.

It also makes sure to learn the lemma which is important for making progress. To do this it uses a relation view of the main transition system. It also splits the constraints into the conjunctive partition for a better unsat core.

This PR also fixes an assertion failure that would show up occasionally (more often with mathsat) where the predecessor was over-generalized. There were several iterations to fix this including a recursive justify coi implementation (I think the important part is traversing as a tree instead of a DAG because location of the node matters), and various other attempts. In the end, the only thing that worked was adding all variables appearing in constraints (and in the state update of variables from next constraints). These fixes may be overly conservative, but at least they work and can be improved upon in a separate PR. There is good evidence that this assertion failure no longer occurs (several runs with different random seeds don't turn it up). Unfortunately this is highly non-deterministic so I can't be fully sure.

There are still some expected errors when using mathsat as the underlying solver. There appears to be a rewriting bug so that when we create an equality that should be guaranteed to be true in the model (because they had the same value) it ends up being a different term (e.g. rewritten to one that evaluates to false). This bug was already reported and there's not much to do now. However, this seems okay because boolector is better for this algorithm anyway.

Commits:

* deduplicate labels

* comment

* Fix in test

* Add a new test

* Add a fallback refinement procedure

* Straightforward substitution

* Use relational system in IC3SA for learning path constraints

* Minor

* Learn path axiom

* Value substitution with one input unrolled (state var replaced)

* Have two refinement options

* Add ic3sa to ic3_variants

* Get all possible controlling terms

* recursive justify_coi

* add constraints

* Get a reference

* Add learned lemma to trans_label_

* Justify coi for all constraints

* Justify next constraint and add a comment

* Return new terms when updating abstraction and only add symbols from new terms

* Just include all variables from constraints

* Fix state update lookup