BMC option: allow non-minimal cex (#303)
Add option "--bmc-allow-non-minimal-cex" to BMC to skip the search for the minimal counterexample within an interval. This is relevant only when option "--bmc-single-bad-state" is not given and when combined with option "--bmc-bound step <n>" where <n> is greater than 1. Unlike "--bmc-single-bad-state",  option "--bmc-allow-non-minimal-cex" guarantees that a (non-minimal) counterexample can be found, if one exists