New options and improvements to interpolation-based engine (#381)
* Add 3 new options to interp engine

The new options are currently only being passed to the prover,
but have no effect on the solving procedures.

* Ignore .vscode folder

* Implement frontier set simplification for interpolation-based model checking

This optimization is described in Sect. 3.2 of the original paper
"Interpolation and SAT-Based Model Checking" (https://doi.org/10.1007/978-3-540-45069-6_1).

* Skip properties at intermediate time frames when computing interpolants

This optimization is described in Sect. 3.2 of the original paper
"Interpolation and SAT-Based Model Checking" (https://doi.org/10.1007/978-3-540-45069-6_1),
with slight modification that the property at the 1st time frame is also considered.
Otherwise, the computed interpolant (and hence the fixed point) might contain (unreachable) bad states.

* Unroll the transition system eagerly in interp engine

For each computed interpolant, we can extend the safe bound by one.

* Add an option to compute "backward" interpolants

A forward interpolant I of formulas A and B can be computed as:
  I = get_interpolant(A, B)
assuming A ∧ B is unsat.

A backward interpolant I' can be computed as:
  I' = ¬ get_interpolant(B, A)

I' still meets all the requirements of a valid Craig interpolant.

* Enable cvc5's interpolation

* Add more configurable options on which time frames properties are considered during interpolation
- see also 6903e6dc78418a2b56d6176c1f2892cdceaea104

* Fix arg parsing for --interp-props

* Enable interp's frontier-set simplication by default

* Revert "Enable cvc5's interpolation"

This reverts commit f7403ed4be6b3675241b3ee36a073446b3771170,
which will be added to another branch.

* Break correctly when parsing options

* Improve help message for --interp-backward

* Extend notes on interp's eager unrolling

* Add tests for new interp options

* Remove option "only-last" from --interp-props

* Improve naming of enums

* Fix tests following c82afe0 and e3794b4

---------

Co-authored-by: Ahmed Irfan <43099566+ahmed-irfan@users.noreply.github.com>